@@ -21,55 +21,52 @@
     "@angular/core": "^16.2.12",
     "@angular/elements": "^16.2.12",
     "@angular/forms": "^16.2.12",
-    "@angular/google-maps": "^13.3.3",
     "@angular/localize": "^16.2.12",
-    "@angular/material": "^15.2.9",
+    "@angular/material": "^16.2.14",
     "@angular/platform-browser": "^16.2.12",
     "@angular/platform-browser-dynamic": "^16.2.12",
     "@angular/router": "^16.2.12",
-    "@ng-bootstrap/ng-bootstrap": "12.0.1",
+    "@ng-bootstrap/ng-bootstrap": "^15.1.2",
+    "@ngrx/effects": "^16.3.0",
+    "@ngrx/store": "^16.3.0",
+    "@ngrx/store-devtools": "^16.3.0",
     "@popperjs/core": "^2.11.4",
-    "ajv": "7.0.1",
-    "ajv-keywords": "4.0.0",
-    "bootstrap": "4.5.3",
-    "chart.js": "2.9.4",
-    "chartist": "0.11.4",
+    "ajv": "8.12.0",
+    "ajv-keywords": "5.1.0",
+    "bootstrap": "5.3.2",
     "classlist.js": "1.1.20150312",
-    "copy-webpack-plugin": "7.0.0",
-    "core-js": "3.8.1",
+    "copy-webpack-plugin": "12.0.2",
+    "core-js": "3.36.0",
     "crypto-js": "^4.1.1",
     "dom-to-image": "^2.6.0",
     "font-awesome": "4.7.0",
-    "jquery": "3.5.1",
+    "jquery": "3.6.1",
     "jspdf": "^2.5.1",
     "jw-bootstrap-switch-ng2": "2.0.5",
-    "ng2-charts": "2.4.2",
-    "ngx-toastr": "13.2.0",
     "nouislider": "14.6.3",
-    "perfect-scrollbar": "1.5.0",
+    "perfect-scrollbar": "1.5.5",
     "popper.js": "1.16.1",
     "rellax": "1.12.1",
-    "rxjs": "^6.5.4",
+    "rxjs": "~7.8.0",
     "rxjs-compat": "6.6.7",
     "web-animations-js": "2.3.2",
-    "zone.js": "~0.13.3"
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^16.2.12",
     "@angular/cli": "^16.2.12",
     "@angular/compiler-cli": "^16.2.12",
     "@angular/language-service": "16.2.12",
-    "@types/bootstrap": "5.0.1",
-    "@types/chartist": "0.11.0",
+    "@types/bootstrap": "5.2.10",
     "@types/crypto-js": "^4.1.1",
     "@types/dom-to-image": "^2.6.4",
-    "@types/googlemaps": "3.40.5",
+    "@types/googlemaps": "3.43.3",
     "@types/jasmine": "~3.10.3",
     "@types/jasminewd2": "~2.0.10",
-    "@types/jquery": "3.5.5",
-    "@types/node": "^17.0.21",
+    "@types/jquery": "3.5.29",
+    "@types/node": "^20.11.19",
     "codelyzer": "6.0.2",
-    "jasmine-core": "~4.0.1",
+    "jasmine-core": "~5.1.2",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "^6.3.17",
     "karma-chrome-launcher": "~3.1.1",
@@ -79,6 +76,6 @@
     "karma-jasmine-html-reporter": "^1.7.0",
     "protractor": "7.0.0",
     "ts-node": "~10.7.0",
-    "typescript": "~5.1.6"
+    "typescript": "<5.2"
   }
 }
