@@ -1,13 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import {Destinations} from "../../shared/models/destinations";
-import {DestinationsService} from "../../shared/services/api/destinations/destinations.service";
 import {Router} from "@angular/router";
 import {NgForm} from "@angular/forms";
 import {Colis} from "../../shared/models/colis";
-import {ColisService} from "../../shared/services/api/colis/colis.service";
-import {ProfileService} from "../../shared/services/profile/profile.service";
-import {ApiUserService} from "../../shared/services/api/user/api-user.service";
-import {User} from "../../shared/models/user";
+import {DestinationsService} from "../../core/services/destinations.service";
+import {ColisService} from "../../core/services/colis.service";
 
 @Component({
   selector: 'app-colis-new',
@@ -21,18 +18,15 @@ export class ColisNewComponent implements OnInit {
 
 
   wrongCredential: boolean;
-  uniqDest: boolean;
 
 
   constructor(
     private destinationsService: DestinationsService,
-    private router: Router,
-    private colisService: ColisService,
-    private profilService:ProfileService,
-    private userService:ApiUserService,
+    // private router: Router,
+    // private colisService: ColisService,
   ) { }
 
-  user
+  user: any
 
   ngOnInit(): void {
     this.colis.status = 'ATTENTE_EXPEDITION'
@@ -44,7 +38,7 @@ export class ColisNewComponent implements OnInit {
     this.colis.prixTotal = 0
     this.destinationsService.getDestinations().subscribe( result => {
       this.destinations = result
-      this.colis.destination = "/api/destinations/"+result[0].id
+      this.colis.destination = "/api/destinations/"+result[0]._id
       this.colis.prixKilo = result[0].prixKilos
       this.colis.douane = result[0].prixDouane
       this.colis.paysDestination = result[0].pays
@@ -52,10 +46,10 @@ export class ColisNewComponent implements OnInit {
   }
 
   newColi(form: NgForm) {
-    this.profilService.getMe().subscribe(user => {
-      this.user = user
-      this.colis.employe = '/api/users/'+this.user._id
-    })
+    // this.profilService.getMe().subscribe(user => {
+    //   this.user = user
+    //   this.colis.employe = '/api/users/'+this.user._id
+    // })
 
     this.wrongCredential = false
     this.colis.prixTotal = (this.colis.prixKilo*this.colis.poids)+this.colis.douane+this.colis.emballage
@@ -71,30 +65,28 @@ export class ColisNewComponent implements OnInit {
       this.colis.reste = this.colis.prixTotal - this.colis.avance
     }
 
-    this.colisService.findColisByCode(this.colis.numero).subscribe(colF => {
-      let coliFind = colF[0]
-      if (coliFind){
-        this.router.navigate(['/nouveau-coli'])
-      }else {
-        this.colisService.postColi(this.colis).subscribe( resuslt => {
-          this.router.navigate(['/colis'])
-        },error => {
-          this.wrongCredential = true
-        })
-      }
-    })
+    // this.colisService.findColisByCode(this.colis.numero).subscribe(colF => {
+    //   let coliFind = colF[0]
+    //   if (coliFind){
+    //     this.router.navigate(['/nouveau-coli'])
+    //   }else {
+    //     this.colisService.postColi(this.colis).subscribe( resuslt => {
+    //       this.router.navigate(['/colis'])
+    //     },error => {
+    //       this.wrongCredential = true
+    //     })
+    //   }
+    // })
   }
 
 
   changeDestination() {
-    this.destinationsService.getDestinationByURI(this.colis.destination).subscribe( result => {
-      this.colis.prixKilo = result.prixKilos
-      this.colis.paysDestination = result.pays
-      this.colis.douane = result.prixDouane
-    },error => {
-      this.wrongCredential = true
-    })
-
-
+    // this.destinationsService.getDestinationByURI(this.colis.destination).subscribe( result => {
+    //   this.colis.prixKilo = result.prixKilos
+    //   this.colis.paysDestination = result.pays
+    //   this.colis.douane = result.prixDouane
+    // },error => {
+    //   this.wrongCredential = true
+    // })
   }
 }
