@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import {DestinationsService} from "../../shared/services/api/destinations/destinations.service";
 import {Router} from "@angular/router";
 import {Destinations} from "../../shared/models/destinations";
+import {DestinationsService} from "../../core/services/destinations.service";
 
 @Component({
   selector: 'app-destinationss',
@@ -28,12 +28,12 @@ export class DestinationsComponent implements OnInit {
 
   lockUser(id: number, b: boolean, $event: MouseEvent) {
     let destination: Destinations = new Destinations()
-    destination.id = id
+    destination._id = id
     destination.isArchivate = b
-    this.destinationsService.deleteDestinationr(destination).subscribe( resuslt => {
-      this.getDestinations()
-    },error => {
-
-    })
+    // this.destinationsService.deleteDestinationr(destination).subscribe( resuslt => {
+    //   this.getDestinations()
+    // },error => {
+    //
+    // })
   }
 }
