@@ -1,55 +1,62 @@
 import { Component, OnInit } from '@angular/core';
-import {Destinations} from "../../shared/models/destinations";
-import {DestinationsService} from "../../shared/services/api/destinations/destinations.service";
+// import {Destinations} from "../../shared/models/destinations";
+// import {DestinationsService} from "../../shared/services/api/destinations/destinations.service";
 import {ActivatedRoute, Router} from "@angular/router";
 import {NgForm} from "@angular/forms";
 import {Pays} from "../../shared/models/pays";
-import {PaysService} from "../../shared/services/api/pays/pays.service";
+import {ModuleStoreService} from "../../core/store/module-store.service";
+// import {PaysService} from "../../shared/services/api/pays/pays.service";
 
 @Component({
   selector: 'app-pays-details',
   templateUrl: './pays-details.component.html',
   styleUrls: ['./pays-details.component.scss']
 })
-export class PaysDetailsComponent implements OnInit {
+export class PaysDetailsComponent{
 
-  pays : Pays
+  pays : Pays = { nom: '', _id: ''}
 
 
   wrongCredential: boolean;
   uniqDest: boolean;
+  id = this.route.snapshot.params['id']
 
   constructor(
-    private paysService: PaysService,
+    private moduleStoreService: ModuleStoreService,
     private router: Router,
     private route:ActivatedRoute,
-  ) { }
-
-  private id = this.route.snapshot.params['id']
-
-
-  ngOnInit(): void {
-    this.paysService.getPays(this.id).subscribe(dest => this.pays = dest)
+  ) {
+    let trouve= false
+    this.moduleStoreService.selectPays().subscribe(pays => {
+      pays.forEach(pay => {
+        if (pay._id == this.id) {
+          this.pays = pay
+          trouve = true
+        }
+      })
+    })
+    if (!trouve) {
+      this.router.navigate(['/pays'])
+    }
   }
+
 
   upDest(form: NgForm) {
     this.wrongCredential = false
-
-
-
     this.uniqDest = false
 
-    this.paysService.findPaysByLibelle(this.pays.nom).subscribe(destF => {
-      let destinationFind = destF[0]
-      if (destinationFind){
-        this.uniqDest = true
+    this.moduleStoreService.selectPays().subscribe(pays => {
+      pays.forEach(pay => {
+        if (pay.nom == this.pays.nom && pay._id != this.pays._id) {
+          this.uniqDest = true
+        }
+      })
+
+      if (this.uniqDest) {
         document.getElementById('nom').classList.add('is-invalid')
-      }else {
-        this.paysService.putPays(this.pays).subscribe( resuslt => {
-          this.router.navigate(['/pays'])
-        },error => {
-          this.wrongCredential = true
-        })
+      } else {
+        this.moduleStoreService.updatePays(this.pays)
+        this.router.navigate(['/pays'])
       }
     })
   }
