@@ -1,38 +1,24 @@
 import { Component, OnInit } from '@angular/core';
 import {User} from "../../shared/models/user";
-import {ApiUserService} from "../../shared/services/api/user/api-user.service";
-import {Route, Router} from "@angular/router";
+import {ModuleStoreService} from "../../core/store/module-store.service";
 
 @Component({
   selector: 'app-users-list',
   templateUrl: './users-list.component.html',
   styleUrls: ['./users-list.component.scss']
 })
-export class UsersListComponent implements OnInit {
+export class UsersListComponent {
+
+  $users = this.moduleStoreService.selectEmployees()
 
   constructor(
-    private userService:ApiUserService,
-    private router: Router
-  ) { }
-
-  users:User[]=[]
-  getUsers(){
-    this.userService.getUsers().subscribe(res => {
-      this.users = res
-    })
-  }
-  ngOnInit(): void {
-    this.getUsers()
+    private moduleStoreService: ModuleStoreService,
+  ) {
   }
 
-  lockUser(id: number, b: boolean, $event: MouseEvent) {
-    let user: User
-    user.isLocked = b
-    user._id = id
-    this.userService.putUser(user).subscribe( resuslt => {
-      this.getUsers()
-    },error => {
 
-    })
+  lockUser(user: User, lock: boolean) {
+    user.isLocked = lock
+    this.moduleStoreService.updateEmployee(user)
   }
 }
